fix imports in Assignment class, create test cases
